packages/decorators/src/decorators/Positive.spec.ts
const addConstraint: jest.Mock<any> = jest.fn();
jest.setMock('./addConstraint', {addConstraint: addConstraint});
import {
    Positive,
    PositiveOrZero
} from './Positive';

describe('@Positive decorator', () => {

    it('adds a constraint', () => {
        // given
        const target: object|Function = function(): void { /* empty */ };
        const propertyKey: string|symbol = 'test';
        addConstraint.mockReset();
        // when
        Positive(target, propertyKey, undefined);
        // then
        expect(addConstraint).toHaveBeenCalledTimes(1);
        expect(addConstraint).toHaveBeenCalledWith({target, propertyKey, descriptor: undefined, constraintName: 'Positive', attributes: {zero: false}});
    });

});

describe('@PositiveOrZero decorator', () => {

    it('adds a constraint', () => {
        // given
        const target: object|Function = function(): void { /* empty */ };
        const propertyKey: string|symbol = 'test';
        addConstraint.mockReset();
        // when
        PositiveOrZero(target, propertyKey, undefined);
        // then
        expect(addConstraint).toHaveBeenCalledTimes(1);
        expect(addConstraint).toHaveBeenCalledWith({target, propertyKey, descriptor: undefined, constraintName: 'Positive', decoratorName: 'PositiveOrZero', attributes: {zero: true}});
    });

});
